cleanup inference code
