+++ inference.py
@@ -10,9 +10,8 @@ import torch
 from torchvision.transforms import transforms
 from pytorch_lightning import LightningModule
 from src.utils import Calib, get_logger
-from src.datamodules.components.kitti_dataset import DetectedObject
 from src.utils.ClassAverages import ClassAverages
-from src.utils.Plotting import plot_3d_box
+from src.utils.Plotting import calc_alpha, plot_3d_box
 from src.utils.Math import calc_location
 from src.utils.Plotting import calc_theta_ray
 
@@ -35,9 +34,6 @@ def inference(config: DictConfig):
 
     # Averages Dimension list
     class_averages = ClassAverages()
-
-    # angle bins
-    angle_bins = generate_bins(bins=2)
 
     # init detector model
     log.info(f"Instantiating detector <{config.detector._target_}>")
@@ -85,39 +81,30 @@ def inference(config: DictConfig):
                 DIMS.append(dim)
             except:
                 dim = DIMS[-1]
-            # calculate theta ray
+            # calculate orientation
             theta_ray = calc_theta_ray(img.size[0], det["box"], proj_matrix)
-            # box_2d
-            box_xyxy = [x.numpy() for x in det['box']]
-            # TODO: understand this
-            argmax = np.argmax(conf)
-            orient = orient[argmax, :]
-            cos = orient[0]
-            sin = orient[1]
-            alpha = np.arctan2(sin, cos)
-            alpha += angle_bins[argmax]
-            alpha -= np.pi
+            alpha = calc_alpha(orient=orient, conf=conf, bins=2)
+            orient = alpha + theta_ray
             # calculate the location
             location, x = calc_location(
                 dimension=dim,
                 proj_matrix=proj_matrix,
-                box_2d=box_xyxy,
+                box_2d=[x.numpy() for x in det["box"]],
                 alpha=alpha,
-                theta_ray=theta_ray
+                theta_ray=theta_ray,
             )
-            # orientation
-            orient = alpha + theta_ray
             # plot 3d bbox
             plot_3d_box(
-                img=img_draw, 
-                cam_to_img=proj_matrix, 
-                ry=orient, 
-                dimension=dim, 
-                center=location
+                img=img_draw,
+                cam_to_img=proj_matrix,
+                ry=orient,
+                dimension=dim,
+                center=location,
             )
 
         if config.get("save_result"):
             cv2.imwrite(f'{config.get("dump_dir")}/{name}_{i:03d}.png', img_draw)
+
 
 def detector_yolov5(model_path: str, cfg_path: str, classes: int, device: str):
     """YOLOv5 detector model"""
@@ -151,15 +138,6 @@ def class_to_labels(class_: int, list_labels: List = None):
         list_labels = ["pedestrian", "car", "truck"]
 
     return list_labels[int(class_)]
-
-def generate_bins(bins):
-    angle_bins = np.zeros(bins)
-    interval = 2 * np.pi / bins
-    for i in range(1, bins):
-        angle_bins[i] = i * interval
-    angle_bins += interval / 2 # center of bins
-
-    return angle_bins
 
 
 if __name__ == "__main__":
